modify likelihood function and random initialization
we increase the max distance and we modify the likelihood. Now we define max_dist as dim/4 so only 1 mode can contribute to the likelihood. In the border, when no mode contributes to the likelihood we define a specific value for the likelihood so the change is smoot enouugh for the bound not to grow to much